requirements.txt
absl-py==2.1.0
altair==5.3.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DAWG-Python==0.7.2
docopt==0.6.2
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.35.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
